// src/interfaces/base.interface.ts
export type SizeType = 'small' | 'medium' | 'large'
export type ColorType =
  | 'primary'
  | 'secondary'
  | 'error'
  | 'warning'
  | 'info'
  | 'success'

export interface BreakpointSizesType {
  xs?: number
  xs2?: number
  sm?: number
  sm2?: number
  md?: number
  md2?: number
  lg?: number
  lg2?: number
  xl?: number
  xl2?: number
  xl3?: number
  xl4?: number
  xl5?: number
}